MP: Add driveForward, driveBackward w/ Task State and isPositionInitialized
The MotionPlanner offers the functionality to drive a given distance straight forward or straight backward.
The Task class contains the corresponding states (BACK_OFF and APPROACH) and the TaskHandler handles the BACK_OFF state correctly.
The APPROACH state is not implemented yet.

In addition the MotionPlanner offers the isPositionInitialized functionality, which will return true, as soon as the poseCallback calls the MotionPlanners update() function the first time. This prevents Path requests from (0.0, 0.0) to somewhere and makes the check for Position (0.0, 0.0) unnecessary.