Roaming: Fix increasing grid nodes and add grid visualization
Robots are allowed to create nodes when they are not exactly on an existing one. This can cause lags when too many nodes are created. Therefore robots always take the clostest node to them.
In addition the grid is now also displayed in free roaming, but can be disabled in the "Display Views" section of rviz for each robot.